+++ components/Code/setup.ts
@@ -9,21 +9,25 @@
  *
  * @module
  */
-import { apDef, o, part } from '../../src/utils.ts';
+import { apDef, forward, o } from '../../src/utils.ts';
 import { iComponent, iFwd } from '../../src/types.ts';
 import { styles } from './styles.ts';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 /** Properties of the `<Code />` component. */
 export type iCode = iComponent & {
-  fwd: Partial<{
-    wrapper: iFwd<HTMLDivElement>;
-  }>;
+  fwd: Partial<iCodeFwd>;
+};
+
+type iCodeFwd = {
+  wrapper: iFwd<HTMLDivElement>;
 };
 
 /** Default values of the `<Code />` component's props. */
 const defaults: iCode = {
-  fwd: {},
+  fwd: {
+    wrapper: {},
+  },
 };
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -31,13 +35,10 @@ const defaults: iCode = {
 export default (props: Partial<iCode>) => {
   const p = apDef<iCode>(defaults, props);
 
-  const { wrapper } = p.fwd;
+  p.class = o('code', { ...p });
+  p.fwd = forward({
+    wrapper: [styles, 'code__wrapper'],
+  }, p.fwd);
 
-  const classes = part({
-    code: o(['code', p.class], { ...p }),
-    wrapper: o(['code__wrapper', styles, wrapper?.class], { ...wrapper }),
-  });
-
-  delete p.class;
-  return { c: classes, ...p };
+  return p;
 };