+++ components/Main/setup.ts
@@ -9,37 +9,18 @@
  *
  * @module
  */
-import { apDef, o, part } from '../../src/utils.ts';
+import { o } from '../../src/utils.ts';
 import { iComponent } from '../../src/types.ts';
-import { LayoutTypes } from '../../src/enums.ts';
-import { iLayout } from '../Layout/setup.ts';
 import { styles } from './styles.ts';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 /** Properties of the `<Main />` component. */
-export type iMain = Omit<iComponent, 'layout'> & {
-  layout: LayoutTypes;
-  fwd: Partial<{
-    layout: Partial<iLayout>;
-  }>;
-};
-
-/** These are the default values of the `<Main />` component's props. */
-const defaults: iMain = {
-  layout: 'default',
-  fwd: {},
-};
+export type iMain = iComponent;
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 /** Setup function of the `<Main />` component. */
 export default (props: Partial<iMain>) => {
-  const p = apDef<iMain>(defaults, props);
+  props.class = o(styles, { ...props });
 
-  const classes = part({
-    main: o(styles, { ...p }),
-    layout: o('', { ...p.fwd.layout }),
-  });
-
-  delete p.class;
-  return { c: classes, ...p };
+  return props;
 };