+++ components/Separator/setup.ts
@@ -9,7 +9,7 @@
  *
  * @module
  */
-import { o, part } from '../../src/utils.ts';
+import { o } from '../../src/utils.ts';
 import { iComponent } from '../../src/types.ts';
 import { styles } from './styles.ts';
 
@@ -20,12 +20,7 @@ export type iSeparator = iComponent<HTMLHRElement>;
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 /** Setup function of the `<Separator />` component. */
 export default (props: Partial<iSeparator>) => {
-  const p = props;
+  props.class = o(styles, { ...props });
 
-  const classes = part({
-    separator: o(styles, { ...p }),
-  });
-
-  delete p.class;
-  return { c: classes, ...p };
+  return props;
 };